Do not cache dependencies due to missing mix.lock
Without this commit, the following error happens in the CI:

```
the dependency is not locked. To generate the "mix.lock" file run "mix deps.get"
```

This is now fixed. The issue was that the cache is restored, but
`mix.lock` is missing since `mix deps.get` wasn't executed due to the
cache hit. `mix.lock` isn't tracked since the Oban Pro dependencies are
only available in a private Hex repository. So for contributors without
the Oban Pro license, fetching those dependencies pinned in `mix.lock`
would be impossible. This is why `mix.lock` is always regenerated.